requirements.txt
aiogram==3.11.0
yadisk==1.4.0
python-dotenv==1.0.1
aiohttp==3.10.0